Set TF_PLUGIN_CACHE_DIR to a path relative to the repo root
The previous implementation was a path relative to the current directory
of tests, but this would result in duplicate copies being scattered
around if imported from other packages. By resolving a path relative to
the file in which the helper function is defined, we allow multiple
packages to share the cache in go test.