Fix CI fail for TestExecutorDir in ubuntu-20.04
The `ubuntu-latest` in GitHub Action has been changed from
`ubuntu-18.04` to `ubuntu-20.04`, In `ubuntu-20.04`, `/bin` is a symlink
to `/usr/bin`, this breaks the test for `TestExecutorDir`. It's ok
changing it to any dir except not a symlink in all platforms on CI. So
we change the test dir to `/usr/local/bin`.