Move storage implementations to a new package
It is natural for most of users to expect the history storage
implementation of tfmigrate to have the same authentication settings and
priorities as the Terraform core backend. To do this, we need to reuse
the code from the upstream.

The problem is that the license of hashicorp/terraform is the MPL2, but
this project is currently distributed under the terms of MIT. As you
know, the MPL2 cannot be redistributed as the MIT.

Before expanding support for other storage types, we are planning to
split the storage-related implementations into a new separate
repository.

As the first step, I'll separate it at the package level.