Copy the storage package from minamijoyo/tfmigrate
https://github.com/minamijoyo/tfmigrate/tree/49ea331bbf97b9effe3344428db52d0f29719800/storage

It's natural that tfmigrate users will expect to be authenticated cloud
providers with the same options and precedences as terraform backend.
This was not a problem in the s3 storage (I think) because the
authentication logic for AWS is implemented in an external library
hashicorp/aws-sdk-go-base. However, except for AWS, to expand support
for other history storage types, we need to reuse the upstream code.

The problem is that the license of hashicorp/terraform is the MPL2, but
the tfmigrate is currently distributed under the terms of MIT. Before
expanding support for other storage types, I'll split the storage
implementations into a new separate repository, which will be
distributed as the MPL2.