Update golangci-lint to v1.45.2 and actions to latest
The golangci/golangci-lint-action@v2 implicitly installs Go and its
version was recently upgraded to Go 1.18. It causes an issue for old
golangci-lint. To fix the problem either upgrading
golangci-lint-action@v3 and setup Go 1.17.x explicitly or updating
golangci-lint to the latest v1.45.2. I fixed both and also upgrading
some actions to the latest.