Revert the deprecation of the tfmigrate plan --out=tfplan option
The `tfmigrate plan --out=tfplan` option was initially implemented in
tfmigrate v0.2.7 and deprecated in v0.3.0 because the feature was based
on a bug before Terraform v1.1. This means that the saved plan file is
no longer applicable in Terraform v1.1+.
https://github.com/minamijoyo/tfmigrate/pull/63

Even though, we found the plan file would still be useful for static
analysis such as Conftest.
https://github.com/minamijoyo/tfmigrate/issues/106

Let’s revert the deprecation and clarify it's intended to use only for
static analysis.

Note that the old terraform show command before Terraform v1.1.3 rejects
a plan file generated by `tfmigrate plan —out=tfplan` as stale. It works
after this change.
https://github.com/hashicorp/terraform/pull/30205

Use Terraform v1.1.3+ for this feature. I think this limitation is
acceptable as long as it works with the latest Terraform.