Set TF_CLI_ARGS_apply to --parallelism=1 in sandbox
The acceptance test is failing quite often these days. I was able to
reproduce the problem in the sandbox environment, so I observed the
state of the localstack when it failed, and found that terraform apply
created only one security group when it should have created three.
From observation, although we don’t have complete confidence in the root
cause, it appears that localstack sometimes misses API requests when run
in parallel.
Let's set the number of parallelism to 1 for terraform apply and see if
it improves or not for a while.