Download the providers and generate a cache once before testing
When using TF_PLUGIN_CACHE_DIR, terraform init is not concurrency safe and
acceptance tests frequently fail with checksum mismatch errors.
https://github.com/hashicorp/terraform/issues/25849
To avoid this problem, download the providers and generate a cache once
before testing. I wish we could use the terraform providers mirror command,
but it's only available in v0.13 or later. We use a simple terraform init to
make it work with Terraform v0.12.