Use a local filesystem mirror on CI
Starting with Terraform v1.4, the global plugin cache is ignored on the
first terraform init. This makes caching in CI meaningless.
https://github.com/hashicorp/terraform/pull/32494

To utilize the cache, we use a local filesystem mirror. Strictly
speaking, the mirror is only available in Terraform v0.13+, but it is
hard to compare versions in bash, so we use the mirror unless v0.x.
https://developer.hashicorp.com/terraform/cli/config/config-file#implied-local-mirror-directories