Avoid using terraform init -from-module
Starting from Terraform v1.5-beta, terraform init -from-module emits a
warning for ignoring backend config. I asked about this issue in the
upstream, and it turned out that the option should not use except for
Terragrunt.
https://github.com/hashicorp/terraform/issues/33276

So, I fixed the getting started guide to use simply copying and
initializing instead of using the option.