Change tfmigrate-storage license from MPL2 to MIT
I originally wrote the local and s3 storage implementation for tfmigrate
as the MIT license. Next, someone sent me a patch for gcs storage
implementation https://github.com/minamijoyo/tfmigrate/pull/23, copied
from the upstream terraform backend implementation. I was concerned it
would violate the MPL2 license because the tfmigrate' license is MIT,
and terraform's is MPL2.

So I split the storage implementations
https://github.com/minamijoyo/tfmigrate/pull/80 into a separate
https://github.com/minamijoyo/tfmigrate-storage repository and
relicensed it as MPL2 so that someone could easily copy snippets from
the upstream. However, we could not contact that patch's author and the
initial patch was not merged finally.

Time passed, KengoTODA sent me another patch
https://github.com/minamijoyo/tfmigrate-storage/pull/3 for gcs support
which contained minimal functionality and was not copied from upstream.
This patch has finally been accepted and merged.

That is, I've changed the license of the storage implementation with the
intention that it will contain a derivative work of terraform that can
be easily copied from terraform's backend implementation to extend
various storage types, but this has yet to happen. It's clearly far from
the terraform's implementation at the time of writing.

Recently, HashiCorp announced that it would change the license of its
products, including Terraform, from MPL2 to BSL1.1.
https://www.hashicorp.com/blog/hashicorp-adopts-business-source-license

If, as I originally intended, we allow contributors to copy code from
upstream terraform, the tfmigrate-storage repository must also be
changed to BSL. This means that the tfmigrate source can continue to be
distributed as MIT, but the tfmigrate's binary will be bounded by BSL.
Even though HashiCorp's BSL is a reasonably permissive license with no
restrictions other than HashiCorp's competitors, it is no longer
open-source by OSI definition. I want to continue to keep tfmigrate as
an open-source project.

Technically, it is possible to keep the tfmigrate-storage repository as
MPL2. Still, I no longer have a solid reason to do so; it is just
inconvenient to have a separate repository.

Fortunately, KengoTODA and I are still the only contributors to the
tfmigrate-storage repository, and I got permission from KengoTODA to
re-license his patch as MIT in
https://github.com/minamijoyo/tfmigrate/issues/146.

So, I'm going to change the license of the tfmigrate-storage repository
to MIT and will merge it into the tfmigrate repository.