Mock iam_endpoint explicitly for testing with localstack
Starting from Terraform v1.6, the s3 backend configuration has changed,
and the tfmigrate acceptance test fails. I suspect that the
authentication mechanism may have been changed with the migration to
aws-sdk-go-v2; I have tried a few things and found that I can
authenticate by specifying the `iam_endpoint`. The `iam_endpoint`
argument has been deprecated since Terraformv1.6. It is recommended to
use the newly added `endpoints` argument, though, for running tests on
multiple Terraform versions, the test settings must remain compatible
with the older versions. This is a test-only issue, so I'll use the
deprecated `iam_endpoint` for now. In the future, we should change the
backend configuration depending on the Terraform version.