Parse and detect Terraform CLI implementation type
We are not using it to determine capability now, but we would like to
distinguish which implementation is being used and output it in the log
for future bug reports.

The execPath is inappropriate here because it could be replaced by a
wrapper such as direnv; it must be detected by parsing the output of the
version command.