Restore and deprecate the `to_skip_plan` in a single-state migration
This is a follow-up to #193. Renaming the `to_skip_plan` attribute in a
single-state migration block to `skip_plan` was an incompatible change
to the HCL file's syntax. Since it is technically possible to maintain
backward compatibility, we will continue to accept the old
`to_skip_plan` and log a warning. This attribute will be removed in a
future version. Use `skip_plan` instead.