Allow setting --config via the environment variable TFMIGRATE_CONFIG
In some use cases, you may want to reuse a configuration file. In that
case, specifying it with the `--config` argument would be tedious each
time. Therefore, we have made it possible to set the path to the
configuration file from an environment variable `TFMIGRATE_CONFIG`.

This is related to the discussion in #199.